Creating python_api file and moved classes IterData() and Connect() to python_api file
